requirements.txt
# Core LLM
openai>=1.30.0
anthropic>=0.25.0

# Agent frameworks (optional, templates work without these)
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-community>=0.2.0

# RAG / Embeddings
chromadb>=0.5.0
tiktoken>=0.7.0
faiss-cpu>=1.8.0

# Web search tools
duckduckgo-search>=6.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0

# Document processing
pypdf>=4.2.0
python-docx>=1.1.0
unstructured>=0.13.0

# Utilities
python-dotenv>=1.0.0
pydantic>=2.7.0
rich>=13.7.0
tenacity>=8.3.0
httpx>=0.27.0
